Legend code to separate module.
