+++ lib/graph3d/stylehandlers/BarSizeStyleHandler.js
@@ -10,6 +10,8 @@ function BarSizeStyleHandler() {
   BarParentStyleHandler.call(this);
 
   this.numberOfColumns = 4;
+  this.isColorBar      = false;
+  this.isValueLegend   = false;
 }
 support.derive(BarSizeStyleHandler, BarParentStyleHandler);
 
@@ -35,18 +37,6 @@ function(graph3d, ctx, point, xWidth, yWidth) {
   yWidth = (graph3d.yBarWidth / 2) * minOffset / denom;
 
   this.parent().redrawPoint.call(this, graph3d, ctx, point, xWidth, yWidth);
-};
-
-
-
-/**
- * Override for method in parent class.
- */
-BarSizeStyleHandler.prototype.drawLegend = function(graph3d, ctx) {
-  var dimensions = support.getLegendDimensions(graph3d);
-
-  support.drawSizeLegendBox(graph3d, ctx, dimensions);
-  support.printValueText(graph3d, ctx, dimensions, false);
 };
