+++ lib/graph3d/stylehandlers/BaseStyleHandler.js
@@ -1,5 +1,7 @@
 var support = require ('./Support');
 var Point3d = require('../Point3d');
+var Legend  = require('../Legend');
+
 
 //--------------------------
 // Class BaseStyleHandler
@@ -9,28 +11,13 @@ function BaseStyleHandler() {
   this.numberOfColumns = 3;
 }
 
+
 BaseStyleHandler.prototype.adjustForBarWidth =
   function(graph3d, xRange, yRange) {};
 
 
 BaseStyleHandler.prototype.drawLegend = function(graph3d, ctx, options) {
-  var dimensions = support.getLegendDimensions(graph3d);
-
-  // If options passed, allow these to override the default dimension values.
-  if (options !== undefined) {
-    support.safeCopy(options, dimensions);
-
-    // recalc left in case width changed
-    dimensions.left = dimensions.right - dimensions.width; // px
-  }
-
-  if (this.isColorBar) {
-    support.drawColorBar(graph3d, ctx, dimensions);
-  } else {
-    support.drawSizeLegendBox(graph3d, ctx, dimensions);
-  }
-
-  support.printValueText(graph3d, ctx, dimensions, this.isValueLegend);
+  Legend.drawLegend(graph3d, ctx, options, this.isColorBar, this.isValueLegend);
 };
 
 
@@ -104,5 +91,6 @@ BaseStyleHandler.prototype.dataPointFromXY = function(graph3d, x, y) {
 
   return closestDataPoint;
 };
+
 
 module.exports = BaseStyleHandler;
