+++ lib/graph3d/stylehandlers/SurfaceStyleHandler.js
@@ -85,15 +85,4 @@ SurfaceStyleHandler.prototype.redrawPoint = function(graph3d, ctx, point) {
 };
 
 
-/**
- * Override for method in parent class.
- */
-SurfaceStyleHandler.prototype.drawLegend = function(graph3d, ctx) {
-  var dimensions = support.getLegendDimensions(graph3d);
-
-  support.drawColorBar(graph3d, ctx, dimensions);
-  support.printValueText(graph3d, ctx, dimensions, false);
-};
-
-
 module.exports = SurfaceStyleHandler;
